refactor: reorganize imports to align with backend structure and improve clarity
