--- mineru/api/__init__.py
@@ -1,1 +1,0 @@
-# Copyright (c) Opendatalab. All rights reserved.
